+++ tpl/default.php
@@ -1,21 +1,17 @@
-<style>
-	pre code{
-		display: block;
-	}
-</style>
+<div class="wrap">
+	<h1><?= get_admin_page_title() ?></h1>
 
-<div class="card">
-	<h1>WPHelper\AdminMenuPage</h1>
-	<h2><?php esc_html( get_admin_page_title() ) ?></h2>
-	<p>Please provide a template file or callback function to render this page
-	<br />Like so:
-	</p>
-	<pre><code>
-new AdminMenuPage(
-  [
-    'render' => 'callback_or_tpl_file',
-  ]
-);
-
-</code></pre>
+	<div class="card">
+		<h3>WPHelper\AdminMenuPage</h3>
+		<p>Please provide a template file or callback function to render this page
+		<br />Like so:
+		</p>
+		<pre><code style="display: block;">new WPHelper\AdminMenuPage(
+	[
+		'slug'     => '<?=$this->slug?>',
+		'title'    => '<?=$this->title?>',
+		<strong><em>'render' => 'callback_or_tpl_file',</em></strong>
+	]
+);</code></pre>
+	</div>
 </div>